enable a downloaded pack in the main thread
The download of the pack runs in a different thread. Enabling
the pack may reload some textures and needs an opengl context.
The download thread has no opengl context so enabling the pack
must be moved into the main thread. This ensures the pack
is enabled in the same spot regardless if its already present
or needs to be download.